chiron_server/requirements.txt
# Django
Django==2.0
requests==2.18.4
djangorestframework==3.7.3
djangorestframework-xml==1.3.0
